+++ app/src/screens/Favorites/styles.ts
@@ -1,5 +1,4 @@
-import { ThumbsUp } from 'phosphor-react-native';
-import { RFValue } from 'react-native-responsive-fontsize';
+import { Trash } from 'phosphor-react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 
 import styled, { css } from 'styled-components/native';
@@ -19,6 +18,20 @@ export const Header = styled.View`
 
   margin-bottom: ${({ theme }) => theme.sizes.medium}px;
 `;
+
+export const RemoveButton = styled.TouchableOpacity.attrs(() => ({
+  activeOpacity: 0.7,
+}))`
+  padding: ${({ theme }) => theme.sizes.smallest}px;
+  background-color: ${({ theme }) => theme.colors.gray[400]};
+  border-radius: 9999px;
+`;
+
+export const TrashIcon = styled(Trash).attrs(({ theme }) => ({
+  size: theme.sizes.large,
+  color: theme.colors.red[700],
+  weight: 'light',
+}))``;
 
 export const Title = styled.Text`
   font-size: ${({ theme }) => theme.sizes.xMedium}px;
@@ -47,68 +60,4 @@ export const MovieCard = styled.TouchableOpacity`
     border-radius: ${theme.sizes.small}px;
     margin-bottom: ${theme.sizes.medium}px;
   `}
-`;
-
-export const PictureAndInfo = styled.View`
-  flex: 1;
-
-  flex-direction: row;
-`;
-
-export const MovieInfo = styled.View`
-  flex: 1;
-
-  padding: ${({ theme }) => theme.sizes.medium}px;
-`;
-
-export const Image = styled.Image.attrs(() => ({
-  width: RFValue(80),
-  resizeMode: 'contain',
-}))`
-  height: 100%;
-  border-top-left-radius: ${({ theme }) => theme.sizes.small}px;
-  border-bottom-left-radius: ${({ theme }) => theme.sizes.small}px;
-`;
-
-export const MovieTitle = styled.Text`
-  font-weight: bold;
-  font-size: ${({ theme }) => theme.sizes.medium}px;
-  color: ${({ theme }) => theme.colors.white};
-`;
-
-export const Overview = styled.Text`
-  text-align: justify;
-
-  ${({ theme }) => css`
-    color: ${theme.colors.white};
-    font-size: ${theme.sizes.regular}px;
-    margin-bottom: ${theme.sizes.small}px;
-  `}
-`;
-
-export const VotesContainer = styled.View`
-  flex-direction: row;
-`;
-
-export const VotesText = styled.Text`
-  text-align: justify;
-
-  ${({ theme }) => css`
-    color: ${theme.colors.gray[200]};
-    font-size: ${theme.sizes.regular}px;
-    margin-left: ${theme.sizes.smallest}px;
-  `}
-`;
-
-export const PopularityIcon = styled(ThumbsUp).attrs(({ theme }) => ({
-  size: theme.sizes.medium,
-  color: theme.colors.gray[100],
-  weight: 'bold',
-}))``;
-
-export const TrashContainer = styled.TouchableOpacity`
-  flex-direction: row;
-  position: absolute;
-  right: ${RFValue(6)}px;
-  top: ${RFValue(6)}px;
 `;
